Split SDL2 and SDL1.2 UI for further development Adapt Makefiles
