+++ builds/android/app/src/main/jni/src/Android.mk
@@ -28,7 +28,7 @@ LOCAL_STATIC_LIBRARIES := \
 
 LOCAL_LDLIBS := -lGLESv1_CM -llog -lz -latomic
 
-LOCAL_CFLAGS := -O2 -Wall -Wextra -DUSE_SDL \
+LOCAL_CFLAGS := -O2 -Wall -Wextra -DUSE_SDL=2 \
 		-DHAVE_MPG123 -DWANT_FMMIDI -DHAVE_OPUS -DHAVE_WILDMIDI \
 		-DHAVE_OGGVORBIS -DHAVE_XMP -DHAVE_LIBSNDFILE \
 		-DHAVE_LIBSPEEXDSP -DSUPPORT_AUDIO
