+++ src/sdl_ui.cpp
@@ -20,34 +20,13 @@
 
 #include "system.h"
 
-#ifdef USE_SDL
+#if USE_SDL==1
 
 #include "sdl_ui.h"
 
-#ifdef _WIN32
-#  define WIN32_LEAN_AND_MEAN
-#  ifndef NOMINMAX
-#    define NOMINMAX
-#  endif
-#  include <windows.h>
-#elif defined(GEKKO)
+#if defined(GEKKO)
 #  include <gccore.h>
 #  include <wiiuse/wpad.h>
-#elif defined(__ANDROID__)
-#  include <jni.h>
-#  include <SDL_system.h>
-#elif defined(EMSCRIPTEN)
-#  include <emscripten.h>
-#endif
-#if defined(_WIN32) || SDL_MAJOR_VERSION>1
-	#include "icon.h"
-
-	// Prevent some XLib name clashes under Linux
-	#define Font Font_XLib
-	#define Drawable Drawable_XLib
-	#include "SDL_syswm.h"
-	#undef Font
-	#undef Drawable
 #endif
 
 #include "color.h"
@@ -63,8 +42,6 @@
 
 #  ifdef HAVE_SDL_MIXER
 #    include "audio_sdl_mixer.h"
-#  elif defined(HAVE_OPENAL)
-#    include "audio_al.h"
 #  else
 #    include "audio_sdl.h"
 #  endif
@@ -74,17 +51,10 @@ AudioInterface& SdlUi::GetAudio() {
 }
 #endif
 
-// SDL 1.2 compatibility
-#if SDL_MAJOR_VERSION==1
-	#define SDL_Keycode SDLKey
-	#define SDL_WINDOW_FULLSCREEN_DESKTOP SDL_FULLSCREEN
-	#define SDL_WINDOWEVENT SDL_ACTIVEEVENT
-#endif
-
 static int FilterUntilFocus(const SDL_Event* evnt);
 
 #if defined(USE_KEYBOARD) && defined(SUPPORT_KEYBOARD)
-	static Input::Keys::InputKey SdlKey2InputKey(SDL_Keycode sdlkey);
+	static Input::Keys::InputKey SdlKey2InputKey(SDLKey sdlkey);
 #endif
 
 #if defined(USE_JOYSTICK) && defined(SUPPORT_JOYSTICK)
@@ -111,13 +81,9 @@ SdlUi::SdlUi(long width, long height, bool fs_flag) :
 	SYS_SetResetCallback(GekkoResetCallback);
 #endif
 
-	uint32_t flags = SDL_INIT_VIDEO;
+	uint32_t flags = SDL_INIT_VIDEO|SDL_INIT_TIMER;
 
-#ifndef EMSCRIPTEN
-	flags |= SDL_INIT_TIMER;
-#endif
-
-#if (!defined(NDEBUG) || defined(_WIN32))
+#ifndef NDEBUG
 	flags |= SDL_INIT_NOPARACHUTE;
 #endif
 
@@ -127,22 +93,12 @@ SdlUi::SdlUi(long width, long height, bool fs_flag) :
 	// Set window position to the middle of the screen
 	putenv(const_cast<char *>("SDL_VIDEO_WINDOW_POS=center"));
 #endif
-#ifdef __LINUX__
-	// Set the application class name
-	setenv("SDL_VIDEO_X11_WMCLASS", GAME_TITLE, 0);
-#elif defined(PSP)
-	putenv(const_cast<char *>("SDL_ASPECT_RATIO=4:3"));
-#endif
 
 	if (SDL_Init(flags) < 0) {
 		Output::Error("Couldn't initialize SDL.\n%s\n", SDL_GetError());
 	}
 
-#if SDL_MAJOR_VERSION==1
 	sdl_surface = NULL;
-#else
-	sdl_window = NULL;
-#endif
 
 	BeginDisplayModeChange();
 		if (!RequestVideoMode(width, height, fs_flag)) {
@@ -169,21 +125,13 @@ SdlUi::SdlUi(long width, long height, bool fs_flag) :
 	SDL_JoystickOpen(0);
 #endif
 
-#if defined(USE_MOUSE) && defined(SUPPORT_MOUSE)
-	ShowCursor(true);
-#else
+	// No mouse support
 	ShowCursor(false);
-#endif
 
 #ifdef SUPPORT_AUDIO
 #  ifdef HAVE_SDL_MIXER
 	if (!Player::no_audio_flag) {
 		audio_.reset(new SdlMixerAudio());
-		return;
-	}
-#  elif defined(HAVE_OPENAL)
-	if (!Player::no_audio_flag) {
-		audio_.reset(new ALAudio());
 		return;
 	}
 #  else
@@ -206,13 +154,10 @@ uint32_t SdlUi::GetTicks() const {
 }
 
 void SdlUi::Sleep(uint32_t time) {
-#ifndef EMSCRIPTEN
 	SDL_Delay(time);
-#endif
 }
 
 bool SdlUi::RequestVideoMode(int width, int height, bool fullscreen) {
-#if SDL_MAJOR_VERSION==1
 	// FIXME: Split method into submethods, really, this method isn't nice.
 	// Note to Zhek, don't delete this fixme again.
 	const SDL_VideoInfo *vinfo;
@@ -244,14 +189,10 @@ bool SdlUi::RequestVideoMode(int width, int height, bool fullscreen) {
 				// Set up...
 				current_display_mode.flags = flags;
 
+				// All modes available
 				if (modes == (SDL_Rect **)-1) {
-					// All modes available
 					// If we have a high res, turn zoom on
-// FIXME: Detect SDL version for this. current_h and current_w are only available in >1.2.10
-// PSP SDL port is older than this, lol
-#ifndef PSP
 					current_display_mode.zoom = (vinfo->current_h > height*2 && vinfo->current_w > width*2);
-#endif
 #if defined(SUPPORT_ZOOM)
 					zoom_available = current_display_mode.zoom;
 #else
@@ -330,26 +271,6 @@ bool SdlUi::RequestVideoMode(int width, int height, bool fullscreen) {
 
 	// Didn't find a suitable video mode
 	return false;
-#else
-	// SDL2 documentation says that resolution dependent code should not be used
-	// anymore. The library takes care of it now.
-	current_display_mode.width = width;
-	current_display_mode.height = height;
-	current_display_mode.bpp = 32;
-	if (fullscreen) {
-		current_display_mode.flags |= SDL_WINDOW_FULLSCREEN_DESKTOP;
-	}
-	toggle_fs_available = true;
-
-	current_display_mode.zoom = true;
-#ifdef SUPPORT_ZOOM
-	zoom_available = true;
-#else
-	zoom_available = false;
-#endif
-
-	return true;
-#endif
 }
 
 void SdlUi::BeginDisplayModeChange() {
@@ -396,7 +317,6 @@ bool SdlUi::RefreshDisplayMode() {
 		display_height *= 2;
 	}
 
-#if SDL_MAJOR_VERSION==1
 	int bpp = current_display_mode.bpp;
 
 	// Free non zoomed surface
@@ -419,96 +339,9 @@ bool SdlUi::RefreshDisplayMode() {
 		sdl_surface->format->Bmask,
 		sdl_surface->format->Amask,
 		PF::NoAlpha);
-#else
-	if (!sdl_window) {
-		#ifdef __ANDROID__
-		// Workaround SDL bug: https://bugzilla.libsdl.org/show_bug.cgi?id=2291
-		// Set back buffer format to 565
-		SDL_GL_SetAttribute(SDL_GL_RED_SIZE, 5);
-		SDL_GL_SetAttribute(SDL_GL_GREEN_SIZE, 6);
-		SDL_GL_SetAttribute(SDL_GL_BLUE_SIZE, 5);
-		#endif
-
-		// Create our window
-		sdl_window = SDL_CreateWindow(GAME_TITLE,
-			SDL_WINDOWPOS_CENTERED,
-			SDL_WINDOWPOS_CENTERED,
-			display_width, display_height,
-			SDL_WINDOW_RESIZABLE | flags);
-
-		if (!sdl_window)
-			return false;
-
-		SetAppIcon();
-
-		// OS X needs the rendered to be vsync
-		#if defined(__APPLE__) && defined(__MACH__)
-			uint32_t rendered_flag = SDL_RENDERER_PRESENTVSYNC;
-		#else
-			uint32_t rendered_flag = 0;
-		#endif
-
-		sdl_renderer = SDL_CreateRenderer(sdl_window, -1, rendered_flag);
-		if (!sdl_renderer)
-			return false;
-		SDL_RenderSetLogicalSize(sdl_renderer, SCREEN_TARGET_WIDTH, SCREEN_TARGET_HEIGHT);
-
-		uint32_t const texture_format =
-			SDL_BYTEORDER == SDL_LIL_ENDIAN
-			? SDL_PIXELFORMAT_ABGR8888
-			: SDL_PIXELFORMAT_RGBA8888;
-
-		sdl_texture = SDL_CreateTexture(sdl_renderer,
-			texture_format,
-			SDL_TEXTUREACCESS_STREAMING,
-			SCREEN_TARGET_WIDTH, SCREEN_TARGET_HEIGHT);
-
-		if (!sdl_texture)
-			return false;
-	} else {
-		// Browser handles fast resizing for emscripten, TODO: use fullscreen API
-#ifndef EMSCRIPTEN
-		bool is_fullscreen = (flags & SDL_WINDOW_FULLSCREEN_DESKTOP) == SDL_WINDOW_FULLSCREEN_DESKTOP;
-		if (is_fullscreen) {
-			SDL_SetWindowFullscreen(sdl_window, SDL_WINDOW_FULLSCREEN_DESKTOP);
-		} else {
-			SDL_SetWindowFullscreen(sdl_window, 0);
-			if ((last_display_mode.flags & SDL_WINDOW_FULLSCREEN_DESKTOP)
-					== SDL_WINDOW_FULLSCREEN_DESKTOP) {
-				// Restore to pre-fullscreen size
-				SDL_SetWindowSize(sdl_window, 0, 0);
-			} else {
-				SDL_SetWindowSize(sdl_window, display_width, display_height);
-			}
-		}
-#endif
-	}
-	// Need to set up icon again, some platforms recreate the window when
-	// creating the renderer (i.e. Windows), see also comment in SetAppIcon()
-	SetAppIcon();
-
-	#if SDL_BYTEORDER == SDL_LIL_ENDIAN
-	const DynamicFormat format(
-		32,
-		0x000000FF,
-		0x0000FF00,
-		0x00FF0000,
-		0xFF000000,
-		PF::NoAlpha);
-	#else
-	const DynamicFormat format(
-		32,
-		0xFF000000,
-		0x00FF0000,
-		0x0000FF00,
-		0x000000FF,
-		PF::NoAlpha);
-	#endif
-#endif
 
 	Bitmap::SetFormat(Bitmap::ChooseFormat(format));
 
-#if SDL_MAJOR_VERSION==1
 	if (zoom_available && current_display_mode.zoom) {
 		// Create a non zoomed surface as drawing surface
 		main_surface = Bitmap::Create(current_display_mode.width,
@@ -525,51 +358,27 @@ bool SdlUi::RefreshDisplayMode() {
 		main_surface = Bitmap::Create(
 			pixels, sdl_surface->w, sdl_surface->h, sdl_surface->pitch, format);
 	}
-#else
-	if (!main_surface) {
-		// Drawing surface will be the window itself
-		main_surface = Bitmap::Create(
-			SCREEN_TARGET_WIDTH, SCREEN_TARGET_HEIGHT, Color(0, 0, 0, 255));
-	}
-#endif
 
 	return true;
 }
 
-#ifdef SUPPORT_FULL_SCALING
-void SdlUi::Resize(long width, long height) {
-	if (mode_changing) {
-		current_display_mode.width = width;
-		current_display_mode.height = height;
-	}
-}
-#else
 void SdlUi::Resize(long /*width*/, long /*height*/) {
+	// no-op
 }
-#endif
 
 void SdlUi::ToggleFullscreen() {
 	BeginDisplayModeChange();
 	if (toggle_fs_available && mode_changing) {
-		if ((current_display_mode.flags & SDL_WINDOW_FULLSCREEN_DESKTOP) == SDL_WINDOW_FULLSCREEN_DESKTOP)
-			current_display_mode.flags &= ~SDL_WINDOW_FULLSCREEN_DESKTOP;
+		if ((current_display_mode.flags & SDL_FULLSCREEN) == SDL_FULLSCREEN)
+			current_display_mode.flags &= ~SDL_FULLSCREEN;
 		else
-			current_display_mode.flags |= SDL_WINDOW_FULLSCREEN_DESKTOP;
+			current_display_mode.flags |= SDL_FULLSCREEN;
 	}
 	EndDisplayModeChange();
 }
 
 void SdlUi::ToggleZoom() {
 	BeginDisplayModeChange();
-#if SDL_MAJOR_VERSION > 1
-	// Work around a SDL bug which doesn't demaximize the window when the size
-	// is changed
-	int flags = SDL_GetWindowFlags(sdl_window);
-	if ((flags & SDL_WINDOW_MAXIMIZED) == SDL_WINDOW_MAXIMIZED) {
-		SDL_RestoreWindow(sdl_window);
-	}
-#endif
-
 	if (zoom_available && mode_changing) {
 		current_display_mode.zoom = !current_display_mode.zoom;
 	}
@@ -578,14 +387,6 @@ void SdlUi::ToggleZoom() {
 
 void SdlUi::ProcessEvents() {
 	SDL_Event evnt;
-
-#if defined(USE_MOUSE) && defined(SUPPORT_MOUSE)
-	// Reset Mouse scroll
-	if (Player::mouse_flag) {
-		keys[Input::Keys::MOUSE_SCROLLUP] = false;
-		keys[Input::Keys::MOUSE_SCROLLDOWN] = false;
-	}
-#endif
 
 	// Poll SDL events and process them
 	while (SDL_PollEvent(&evnt)) {
@@ -597,38 +398,22 @@ void SdlUi::ProcessEvents() {
 }
 
 void SdlUi::UpdateDisplay() {
-#if SDL_MAJOR_VERSION==1
 	if (zoom_available && current_display_mode.zoom) {
 		// Blit drawing surface x2 scaled over window surface
 		Blit2X(*main_surface, sdl_surface);
 	}
 	SDL_UpdateRect(sdl_surface, 0, 0, 0, 0);
-#else
-	SDL_UpdateTexture(sdl_texture, NULL, main_surface->pixels(), main_surface->pitch());
-	SDL_RenderClear(sdl_renderer);
-	SDL_RenderCopy(sdl_renderer, sdl_texture, NULL, NULL);
-	SDL_RenderPresent(sdl_renderer);
-#endif
 }
 
 void SdlUi::SetTitle(const std::string &title) {
-#if SDL_MAJOR_VERSION==1
 	SDL_WM_SetCaption(title.c_str(), NULL);
-#else
-	SDL_SetWindowTitle(sdl_window, title.c_str());
-#endif
 }
 
 bool SdlUi::ShowCursor(bool flag) {
-#ifdef __WINRT__
-	// Prevent cursor hide in WinRT because it is hidden everywhere while the app runs...
-	return flag;
-#else
 	bool temp_flag = cursor_visible;
 	cursor_visible = flag;
 	SDL_ShowCursor(flag ? SDL_ENABLE : SDL_DISABLE);
 	return temp_flag;
-#endif
 }
 
 void SdlUi::Blit2X(Bitmap const& src, SDL_Surface* dst_surf) {
@@ -654,7 +439,7 @@ void SdlUi::Blit2X(Bitmap const& src, SDL_Surface* dst_surf) {
 
 void SdlUi::ProcessEvent(SDL_Event &evnt) {
 	switch (evnt.type) {
-		case SDL_WINDOWEVENT:
+		case SDL_ACTIVEEVENT:
 			ProcessActiveEvent(evnt);
 			return;
 
@@ -670,21 +455,6 @@ void SdlUi::ProcessEvent(SDL_Event &evnt) {
 			ProcessKeyUpEvent(evnt);
 			return;
 
-		case SDL_MOUSEMOTION:
-			ProcessMouseMotionEvent(evnt);
-			return;
-
-#if SDL_MAJOR_VERSION>1
-		case SDL_MOUSEWHEEL:
-			ProcessMouseWheelEvent(evnt);
-			return;
-#endif
-
-		case SDL_MOUSEBUTTONDOWN:
-		case SDL_MOUSEBUTTONUP:
-			ProcessMouseButtonEvent(evnt);
-			return;
-
 		case SDL_JOYBUTTONDOWN:
 		case SDL_JOYBUTTONUP:
 			ProcessJoystickButtonEvent(evnt);
@@ -697,40 +467,20 @@ void SdlUi::ProcessEvent(SDL_Event &evnt) {
 		case SDL_JOYAXISMOTION:
 			ProcessJoystickAxisEvent(evnt);
 			return;
-
-#if SDL_MAJOR_VERSION>1
-		case SDL_FINGERDOWN:
-		case SDL_FINGERUP:
-		case SDL_FINGERMOTION:
-			ProcessFingerEvent(evnt);
-			return;
-#endif
 	}
 }
 
 void SdlUi::ProcessActiveEvent(SDL_Event &evnt) {
 	int state;
-#if SDL_MAJOR_VERSION==1
 	state = evnt.active.state;
-#else
-	state = evnt.window.event;
-#endif
-#if PAUSE_GAME_WHEN_FOCUS_LOST
-	if (
-#if SDL_MAJOR_VERSION==1
-	(state == SDL_APPINPUTFOCUS && !evnt.active.gain)
-#else
-	state == SDL_WINDOWEVENT_FOCUS_LOST
-#endif
-	) {
 
+#if PAUSE_GAME_WHEN_FOCUS_LOST
+	if (state == SDL_APPINPUTFOCUS && !evnt.active.gain) {
 		Player::Pause();
 
 		bool last = ShowCursor(true);
 
-#ifndef EMSCRIPTEN
 		// Filter SDL events until focus is regained
-
 		SDL_Event wait_event;
 
 		while (SDL_WaitEvent(&wait_event)) {
@@ -738,7 +488,6 @@ void SdlUi::ProcessActiveEvent(SDL_Event &evnt) {
 				break;
 			}
 		}
-#endif
 
 		ShowCursor(last);
 
@@ -747,20 +496,6 @@ void SdlUi::ProcessActiveEvent(SDL_Event &evnt) {
 
 		return;
 	}
-#endif
-#if defined(USE_MOUSE) && defined(SUPPORT_MOUSE)
-#if SDL_MAJOR_VERSION==1
-	if (state == SDL_APPMOUSEFOCUS) {
-		mouse_focus = evnt.active.gain == 1;
-		return;
-	}
-#else
-	if (state == SDL_WINDOWEVENT_ENTER) {
-		mouse_focus = true;
-	} else if (state == SDL_WINDOWEVENT_LEAVE) {
-		mouse_focus = false;
-	}
-#endif
 #endif
 }
 
@@ -780,11 +515,7 @@ void SdlUi::ProcessKeyDownEvent(SDL_Event &evnt) {
 	}
 
 	// Update key state
-#  if SDL_MAJOR_VERSION==1
 	keys[SdlKey2InputKey(evnt.key.keysym.sym)] = true;
-#  else
-	keys[SdlKey2InputKey(evnt.key.keysym.scancode)] = true;
-#  endif
 #else
 	/* unused */
 	(void) evnt;
@@ -793,76 +524,7 @@ void SdlUi::ProcessKeyDownEvent(SDL_Event &evnt) {
 
 void SdlUi::ProcessKeyUpEvent(SDL_Event &evnt) {
 #if defined(USE_KEYBOARD) && defined(SUPPORT_KEYBOARD)
-#  if SDL_MAJOR_VERSION==1
 	keys[SdlKey2InputKey(evnt.key.keysym.sym)] = false;
-#  else
-	keys[SdlKey2InputKey(evnt.key.keysym.scancode)] = false;
-#  endif
-#else
-	/* unused */
-	(void) evnt;
-#endif
-}
-
-void SdlUi::ProcessMouseMotionEvent(SDL_Event& evnt) {
-#if defined(USE_MOUSE) && defined(SUPPORT_MOUSE)
-	mouse_focus = true;
-	mouse_x = evnt.motion.x;
-	mouse_y = evnt.motion.y;
-#else
-	/* unused */
-	(void) evnt;
-#endif
-}
-
-#if SDL_MAJOR_VERSION>1
-void SdlUi::ProcessMouseWheelEvent(SDL_Event& evnt) {
-#if defined(USE_MOUSE) && defined(SUPPORT_MOUSE)
-	if (!Player::mouse_flag)
-		return;
-
-	// Ignore Finger (touch) events here
-	if (evnt.wheel.which == SDL_TOUCH_MOUSEID)
-		return;
-
-	int amount = evnt.wheel.y;
-
-	// FIXME: Debian ships older SDL2
-#if SDL_VERSION_ATLEAST(2, 0, 4)
-	// translate direction
-	if (evnt.wheel.direction == SDL_MOUSEWHEEL_FLIPPED)
-		amount *= -1;
-#endif
-
-	keys[Input::Keys::MOUSE_SCROLLUP] = amount > 0;
-	keys[Input::Keys::MOUSE_SCROLLDOWN] = amount < 0;
-#else
-	/* unused */
-	(void) evnt;
-#endif
-}
-#endif
-
-void SdlUi::ProcessMouseButtonEvent(SDL_Event& evnt) {
-#if defined(USE_MOUSE) && defined(SUPPORT_MOUSE)
-	if (!Player::mouse_flag)
-		return;
-
-	// Ignore Finger (touch) events here
-	if (evnt.button.which == SDL_TOUCH_MOUSEID)
-		return;
-
-	switch (evnt.button.button) {
-	case SDL_BUTTON_LEFT:
-		keys[Input::Keys::MOUSE_LEFT] = evnt.button.state == SDL_PRESSED;
-		break;
-	case SDL_BUTTON_MIDDLE:
-		keys[Input::Keys::MOUSE_MIDDLE] = evnt.button.state == SDL_PRESSED;
-		break;
-	case SDL_BUTTON_RIGHT:
-		keys[Input::Keys::MOUSE_RIGHT] = evnt.button.state == SDL_PRESSED;
-		break;
-	}
 #else
 	/* unused */
 	(void) evnt;
@@ -945,96 +607,6 @@ void SdlUi::ProcessJoystickAxisEvent(SDL_Event &evnt) {
 #endif
 }
 
-#if SDL_MAJOR_VERSION>1
-void SdlUi::ProcessFingerEvent(SDL_Event& evnt) {
-#if defined(USE_TOUCH) && defined(SUPPORT_TOUCH)
-	SDL_TouchID touchid;
-	int fingers = 0;
-
-	if (!Player::touch_flag)
-		return;
-
-	// We currently ignore swipe gestures
-	if (evnt.type != SDL_FINGERMOTION) {
-		/* FIXME: To simplify things, we lazily only get the current number of
-		   fingers touching the first device (hoping nobody actually uses
-		   multiple devices). This way we do not need to keep track on finger
-		   IDs and deal with the timing.
-		*/
-		touchid = SDL_GetTouchDevice(0);
-		if (touchid != 0)
-			fingers = SDL_GetNumTouchFingers(touchid);
-
-		keys[Input::Keys::ONE_FINGER] = fingers == 1;
-		keys[Input::Keys::TWO_FINGERS] = fingers == 2;
-	}
-#else
-	/* unused */
-	(void) evnt;
-#endif
-}
-
-void SdlUi::SetAppIcon() {
-#if !defined(_WIN32)
-	/* SDL handles transfering the application icon to new or recreated windows,
-	   if initially set through it (see below). So no need to set again for all
-	   platforms relying on it. Platforms defined above need special treatment.
-	*/
-	static bool icon_set = false;
-
-	if (icon_set)
-		return;
-#endif
-	bool load_error = false;
-#ifdef _WIN32
-#ifndef __WINRT__
-	SDL_SysWMinfo wminfo;
-	SDL_VERSION(&wminfo.version)
-	SDL_bool success = SDL_GetWindowWMInfo(sdl_window, &wminfo);
-
-	if (success < 0)
-		Output::Error("Wrong SDL version");
-
-	HWND window;
-	HINSTANCE handle = GetModuleHandle(NULL);
-	HICON icon = LoadIcon(handle, MAKEINTRESOURCE(23456));
-	HICON icon_small = (HICON) LoadImage(handle, MAKEINTRESOURCE(23456), IMAGE_ICON,
-		GetSystemMetrics(SM_CXSMICON), GetSystemMetrics(SM_CYSMICON), 0);
-	load_error = (icon == NULL || icon_small == NULL);
-#endif
-#else
-	//Linux, OS X
-	#if SDL_BYTEORDER == SDL_LIL_ENDIAN
-		uint32_t Rmask = 0x000000FF;
-		uint32_t Gmask = 0x0000FF00;
-		uint32_t Bmask = 0x00FF0000;
-		uint32_t Amask = 0xFF000000;
-	#else
-		uint32_t Rmask = 0xFF000000;
-		uint32_t Gmask = 0x00FF0000;
-		uint32_t Bmask = 0x0000FF00;
-		uint32_t Amask = 0x000000FF;
-	#endif
-	SDL_Surface *icon = SDL_CreateRGBSurfaceFrom(icon32, ICON_SIZE, ICON_SIZE, 32, ICON_SIZE*4, Rmask, Gmask, Bmask, Amask);
-	load_error = (icon == NULL);
-#endif
-	if (load_error)
-		Output::Warning("Could not load window icon.");
-
-#ifdef _WIN32
-#ifndef __WINRT__
-	window = wminfo.info.win.window;
-	SetClassLongPtr(window, GCLP_HICON, (LONG_PTR) icon);
-	SetClassLongPtr(window, GCLP_HICONSM, (LONG_PTR) icon_small);
-#endif
-#else
-	SDL_SetWindowIcon(sdl_window, icon);
-	SDL_FreeSurface(icon);
-	icon_set = true;
-#endif
-}
-#endif
-
 void SdlUi::ResetKeys() {
 	for (size_t i = 0; i < keys.size(); i++) {
 		keys[i] = false;
@@ -1042,13 +614,12 @@ void SdlUi::ResetKeys() {
 }
 
 bool SdlUi::IsFullscreen() {
-	return (current_display_mode.flags & SDL_WINDOW_FULLSCREEN_DESKTOP) == SDL_WINDOW_FULLSCREEN_DESKTOP;
+	return (current_display_mode.flags & SDL_FULLSCREEN) == SDL_FULLSCREEN;
 }
 
 #if defined(USE_KEYBOARD) && defined(SUPPORT_KEYBOARD)
-Input::Keys::InputKey SdlKey2InputKey(SDL_Keycode sdlkey) {
+Input::Keys::InputKey SdlKey2InputKey(SDLKey sdlkey) {
 	switch (sdlkey) {
-#if SDL_MAJOR_VERSION==1
 		case SDLK_BACKSPACE		: return Input::Keys::BACKSPACE;
 		case SDLK_TAB			: return Input::Keys::TAB;
 		case SDLK_CLEAR			: return Input::Keys::CLEAR;
@@ -1143,103 +714,7 @@ Input::Keys::InputKey SdlKey2InputKey(SDL_Keycode sdlkey) {
 		case SDLK_CAPSLOCK		: return Input::Keys::CAPS_LOCK;
 		case SDLK_NUMLOCK		: return Input::Keys::NUM_LOCK;
 		case SDLK_SCROLLOCK		: return Input::Keys::SCROLL_LOCK;
-#else
-		case SDL_SCANCODE_BACKSPACE		: return Input::Keys::BACKSPACE;
-		case SDL_SCANCODE_TAB			: return Input::Keys::TAB;
-		case SDL_SCANCODE_CLEAR			: return Input::Keys::CLEAR;
-		case SDL_SCANCODE_RETURN		: return Input::Keys::RETURN;
-		case SDL_SCANCODE_PAUSE			: return Input::Keys::PAUSE;
-		case SDL_SCANCODE_ESCAPE		: return Input::Keys::ESCAPE;
-		case SDL_SCANCODE_SPACE			: return Input::Keys::SPACE;
-		case SDL_SCANCODE_PAGEUP		: return Input::Keys::PGUP;
-		case SDL_SCANCODE_PAGEDOWN		: return Input::Keys::PGDN;
-		case SDL_SCANCODE_END			: return Input::Keys::ENDS;
-		case SDL_SCANCODE_HOME			: return Input::Keys::HOME;
-		case SDL_SCANCODE_LEFT			: return Input::Keys::LEFT;
-		case SDL_SCANCODE_UP			: return Input::Keys::UP;
-		case SDL_SCANCODE_RIGHT			: return Input::Keys::RIGHT;
-		case SDL_SCANCODE_DOWN			: return Input::Keys::DOWN;
-		case SDL_SCANCODE_PRINTSCREEN	: return Input::Keys::SNAPSHOT;
-		case SDL_SCANCODE_INSERT		: return Input::Keys::INSERT;
-		case SDL_SCANCODE_DELETE		: return Input::Keys::DEL;
-		case SDL_SCANCODE_LSHIFT		: return Input::Keys::LSHIFT;
-		case SDL_SCANCODE_RSHIFT		: return Input::Keys::RSHIFT;
-		case SDL_SCANCODE_LCTRL			: return Input::Keys::LCTRL;
-		case SDL_SCANCODE_RCTRL			: return Input::Keys::RCTRL;
-		case SDL_SCANCODE_LALT			: return Input::Keys::LALT;
-		case SDL_SCANCODE_RALT			: return Input::Keys::RALT;
-		case SDL_SCANCODE_0				: return Input::Keys::N0;
-		case SDL_SCANCODE_1				: return Input::Keys::N1;
-		case SDL_SCANCODE_2				: return Input::Keys::N2;
-		case SDL_SCANCODE_3				: return Input::Keys::N3;
-		case SDL_SCANCODE_4				: return Input::Keys::N4;
-		case SDL_SCANCODE_5				: return Input::Keys::N5;
-		case SDL_SCANCODE_6				: return Input::Keys::N6;
-		case SDL_SCANCODE_7				: return Input::Keys::N7;
-		case SDL_SCANCODE_8				: return Input::Keys::N8;
-		case SDL_SCANCODE_9				: return Input::Keys::N9;
-		case SDL_SCANCODE_A				: return Input::Keys::A;
-		case SDL_SCANCODE_B				: return Input::Keys::B;
-		case SDL_SCANCODE_C				: return Input::Keys::C;
-		case SDL_SCANCODE_D				: return Input::Keys::D;
-		case SDL_SCANCODE_E				: return Input::Keys::E;
-		case SDL_SCANCODE_F				: return Input::Keys::F;
-		case SDL_SCANCODE_G				: return Input::Keys::G;
-		case SDL_SCANCODE_H				: return Input::Keys::H;
-		case SDL_SCANCODE_I				: return Input::Keys::I;
-		case SDL_SCANCODE_J				: return Input::Keys::J;
-		case SDL_SCANCODE_K				: return Input::Keys::K;
-		case SDL_SCANCODE_L				: return Input::Keys::L;
-		case SDL_SCANCODE_M				: return Input::Keys::M;
-		case SDL_SCANCODE_N				: return Input::Keys::N;
-		case SDL_SCANCODE_O				: return Input::Keys::O;
-		case SDL_SCANCODE_P				: return Input::Keys::P;
-		case SDL_SCANCODE_Q				: return Input::Keys::Q;
-		case SDL_SCANCODE_R				: return Input::Keys::R;
-		case SDL_SCANCODE_S				: return Input::Keys::S;
-		case SDL_SCANCODE_T				: return Input::Keys::T;
-		case SDL_SCANCODE_U				: return Input::Keys::U;
-		case SDL_SCANCODE_V				: return Input::Keys::V;
-		case SDL_SCANCODE_W				: return Input::Keys::W;
-		case SDL_SCANCODE_X				: return Input::Keys::X;
-		case SDL_SCANCODE_Y				: return Input::Keys::Y;
-		case SDL_SCANCODE_Z				: return Input::Keys::Z;
-		case SDL_SCANCODE_MENU			: return Input::Keys::MENU;
-		case SDL_SCANCODE_KP_0			: return Input::Keys::KP0;
-		case SDL_SCANCODE_KP_1			: return Input::Keys::KP1;
-		case SDL_SCANCODE_KP_2			: return Input::Keys::KP2;
-		case SDL_SCANCODE_KP_3			: return Input::Keys::KP3;
-		case SDL_SCANCODE_KP_4			: return Input::Keys::KP4;
-		case SDL_SCANCODE_KP_5			: return Input::Keys::KP5;
-		case SDL_SCANCODE_KP_6			: return Input::Keys::KP6;
-		case SDL_SCANCODE_KP_7			: return Input::Keys::KP7;
-		case SDL_SCANCODE_KP_8			: return Input::Keys::KP8;
-		case SDL_SCANCODE_KP_9			: return Input::Keys::KP9;
-		case SDL_SCANCODE_KP_MULTIPLY	: return Input::Keys::MULTIPLY;
-		case SDL_SCANCODE_KP_PLUS		: return Input::Keys::ADD;
-		case SDL_SCANCODE_KP_ENTER		: return Input::Keys::RETURN;
-		case SDL_SCANCODE_KP_MINUS		: return Input::Keys::SUBTRACT;
-		case SDL_SCANCODE_KP_PERIOD		: return Input::Keys::PERIOD;
-		case SDL_SCANCODE_KP_DIVIDE		: return Input::Keys::DIVIDE;
-		case SDL_SCANCODE_F1			: return Input::Keys::F1;
-		case SDL_SCANCODE_F2			: return Input::Keys::F2;
-		case SDL_SCANCODE_F3			: return Input::Keys::F3;
-		case SDL_SCANCODE_F4			: return Input::Keys::F4;
-		case SDL_SCANCODE_F5			: return Input::Keys::F5;
-		case SDL_SCANCODE_F6			: return Input::Keys::F6;
-		case SDL_SCANCODE_F7			: return Input::Keys::F7;
-		case SDL_SCANCODE_F8			: return Input::Keys::F8;
-		case SDL_SCANCODE_F9			: return Input::Keys::F9;
-		case SDL_SCANCODE_F10			: return Input::Keys::F10;
-		case SDL_SCANCODE_F11			: return Input::Keys::F11;
-		case SDL_SCANCODE_F12			: return Input::Keys::F12;
-		case SDL_SCANCODE_CAPSLOCK		: return Input::Keys::CAPS_LOCK;
-		case SDL_SCANCODE_NUMLOCKCLEAR	: return Input::Keys::NUM_LOCK;
-		case SDL_SCANCODE_SCROLLLOCK	: return Input::Keys::SCROLL_LOCK;
-		case SDL_SCANCODE_AC_BACK		: return Input::Keys::AC_BACK;
-		case SDL_SCANCODE_SELECT		: return Input::Keys::SELECT;
 
-#endif
 		default					: return Input::Keys::NONE;
 	}
 }
@@ -1288,22 +763,13 @@ Input::Keys::InputKey SdlJKey2InputKey(int button_index) {
 int FilterUntilFocus(const SDL_Event* evnt) {
 	// Prevent throwing events away received after focus gained but filter
 	// not detached.
-
-	bool filtering_done = false;
-
 	switch (evnt->type) {
 	case SDL_QUIT:
 		Player::exit_flag = true;
 		return 1;
 
-	case SDL_WINDOWEVENT:
-#if SDL_MAJOR_VERSION==1
-		filtering_done = !!(evnt->active.state & SDL_APPINPUTFOCUS);
-#else
-		filtering_done = evnt->window.event == SDL_WINDOWEVENT_FOCUS_GAINED;
-#endif
-
-		return !!filtering_done;
+	case SDL_ACTIVEEVENT:
+		return !!(evnt->active.state & SDL_APPINPUTFOCUS);
 
 	default:
 		return 0;
